requirements.txt
torch>=1.13.0
torchaudio>=2.0.0
pyaudio>=0.2.12
numpy>=1.22.0
whisper>=20230314
silero-vad>=0.4.0
TensorFlowTTS>=0.6
streamlit>=1.22.0
altair>=4.2.0
openai>=0.27.0
gtts>=2.3.1
pydub>=0.25.1
openrouter-ai>=0.1.0
python-dotenv>=1.0.0
requests>=2.28.0
